+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - add `widgets.AccessibleDateInput`
+- update release process
 
 ## [v6.8.0] - 2024-01-17